bixverse meta cell class

This is the bixverse-based meta cell class. It is generated by running meta-cell generating functions on top of the main class. Due to the aggregation of the count matrices to a way smaller number of meta cells, the data is held in memory for this class specifically. Additional methods that make specific usage of meta cells (especially co-expression module detection ones will be made available).

Usage

MetaCells(meta_cell_data, var_data, meta_cell_method)

Arguments

meta_cell_data

Named list. Output of meta-cell generation Rust functions. Will contain the aggregated raw and normalised counts, plus additional information on the origin of the meta cells.

var_data

data.table with the variable/feature informations.

meta_cell_method

String describing the origin of the metacell.

Value

Returns the MetaCells class for further operations.

Properties

obs_table

The meta cell observation table.

var_table

The meta cell variable table.

data

List with the raw and normalised counts.

original_assignment

List with original assignment information.

dims

Dimensions of the new meta cell matrices.

other_data

Potential other data returned from the meta-cell generating methods.