Generates the kNN graph

This function is a wrapper over the Rust-based generation of the approximate nearest neighbours. You have two options to generate the kNNs. "annoy" or "hnsw".

Usage

rs_sc_knn(
  embd,
  no_neighbours,
  n_trees,
  search_budget,
  algorithm_type,
  ann_dist,
  verbose,
  seed
)

Arguments

embd

Numerical matrix. The embedding matrix to use to generate the kNN graph.

no_neighbours

Integer. Number of neighbours to return

n_trees

Integer. Number of trees to use for the "annoy" algorithm.

search_budget

Integer. Search budget per tree for the "annoy" algorithm.

algorithm_type

String. Which of the two implemented algorithms to use. One of c("annoy", "hnsw").

ann_dist

String. The distance metric to use the approximate nearest neighbour search. One of c("cosine", "euclidean").

verbose

Boolean. Controls verbosity of the function and returns how long certain operations took.

seed

Integer. Seed for reproducibility purposes.

Value

A integer matrix of N x k with N being the number of cells and k the number of neighbours.